+++ pytket/extensions/quantinuum/backends/quantinuum.py
@@ -26,6 +26,8 @@ import re
 from typing import Dict, List, Set, Optional, Sequence, Union, Any, cast, Tuple
 from uuid import uuid1
 import warnings
+import datetime
+import zoneinfo
 
 import numpy as np
 import requests
@@ -88,11 +90,7 @@ _STATUS_MAP = {
     "canceled": StatusEnum.CANCELLED,
 }
 
-_GATE_SET = {
-    OpType.Rz,
-    OpType.PhasedX,
-    OpType.ZZMax,
-    OpType.ZZPhase,
+_ADDITIONAL_GATES = {
     OpType.Reset,
     OpType.Measure,
     OpType.Barrier,
@@ -106,6 +104,18 @@ _GATE_SET = {
     OpType.WASM,
 }
 
+_GATE_MAP = {
+    "Rxxyyzz": OpType.TK2,
+    "Rz": OpType.Rz,
+    "RZZ": OpType.ZZPhase,
+    "TK2": OpType.TK2,
+    "U1q": OpType.PhasedX,
+    "ZZ": OpType.ZZMax,
+}
+
+_ALL_GATES = _ADDITIONAL_GATES.copy()
+_ALL_GATES.update(_GATE_MAP.values())
+
 
 def _default_2q_gate(device_name: str) -> OpType:
     # If we change this, we should update the main documentation page and highlight it
@@ -114,9 +124,12 @@ def _default_2q_gate(device_name: str) -> OpType:
 
 
 def _get_gateset(gates: List[str]) -> Set[OpType]:
-    gs = _GATE_SET.copy()
-    if "Rxxyyzz" in gates:
-        gs.add(OpType.TK2)
+    gs = _ADDITIONAL_GATES.copy()
+    for gate in gates:
+        if gate not in _GATE_MAP:
+            warnings.warn(f"Gate {gate} not recognized.")
+        else:
+            gs.add(_GATE_MAP[gate])
     return gs
 
 
@@ -360,6 +373,7 @@ class QuantinuumBackend(Backend):
             dct1["system_type"] = "local_emulator"
             dct1.pop("emulator", None)
             dct1["batching"] = False
+        dct1["cl_reg_width"] = 32 if n_qubits <= 32 else 64
         return BackendInfo(
             name=cls.__name__,
             device_name=name + "LE" if local_emulator else name,
@@ -433,6 +447,96 @@ class QuantinuumBackend(Backend):
         api_handler._response_check(res, "get machine status")
         return str(res.json()["state"])
 
+    def get_calendar(
+        self,
+        start_date: datetime.datetime,
+        end_date: datetime.datetime,
+        localise: bool = True,
+    ) -> List[Dict[str, str]]:
+        r"""Retrieves the Quantinuum H-Series operational calendar
+        for the period specified by start_date and end_date.
+        The calendar data returned is for the local timezone of the
+        end-user.
+
+        The output is a sorted list of dictionaries. Each dictionary is an
+        event on the operational calendar for the period specified by the
+        end-user. The output from this function can be readily used
+        to instantiate a pandas.DataFrame.
+
+        The dictionary has the following properties.
+        * 'start-date': The  start date and start time as a datetime.datetime object.
+        * 'end-date': The end date and end time as a datetime.datetime object.
+        * 'machine': A string specifying the H-Series device attached to the event.
+        * 'event-type': The type of event as a string. The value `online` denotes queued
+            access to the device, and the value `reservation` denotes priority access
+            for a particular organisation.
+        * 'organization': If the 'event-type' is assigned the value 'reservation', the
+            organization with reservation access is specified. Only users within an
+            organization have visibility on organization reservations. Otherwise,
+            organization is listed as 'fairshare', which means all users from all
+            organizations are able to submit jobs to the Fairshare queue during this
+            period.
+
+        :param start_date: The start date as datetime.date object
+            for the period to return the operational calendar.
+        :param end_date: The end date as datetime.date object
+            for the period to return the operational calendar.
+        :param localise: Apply localization to the datetime based
+            on the end-users time zone. Default is True. Disable by
+            setting False.
+        :return: A list of events from the H-Series operational calendar,
+            sorted by the `start-date` of each event. Each event is a python
+            dictionary.
+        :return_type: List[Dict[str, str]]
+        :raises: RuntimeError if an emulator or syntax-checker is specified
+        :raises: ValueError if the argument `start_date` or `end_date` are not
+            datetime.datetime objects.
+        """
+
+        if not isinstance(start_date, datetime.datetime) or not isinstance(
+            end_date, datetime.datetime
+        ):
+            raise ValueError(
+                "start_date and end_date must be datetime.datetime objects."
+            )
+
+        if self._device_name.endswith("E") | self._device_name.endswith("SC"):
+            raise RuntimeError(
+                f"Error requesting data for {self._device_name}. Calendar information not available for emulators (E) or syntax checkers (SC)."
+            )
+
+        l4_calendar_data = self.api_handler.get_calendar(
+            start_date.date().isoformat(), end_date.date().isoformat()
+        )
+        calendar_data = []
+        dt_format = "%a %Y-%m-%d %H:%M (%Z)"
+
+        for l4_event in l4_calendar_data:
+            device_name = l4_event["machine"]
+            if device_name != self._device_name:
+                continue
+            dt_start = _convert_datetime_string(
+                l4_event["start-date"]
+            )  # datetime in UTC tz
+            dt_end = _convert_datetime_string(
+                l4_event["end-date"]
+            )  # datetime in UTC tz
+            if localise:  # Apply timezone localisation on UTC datetime
+                dt_start = dt_start.astimezone()  #
+                dt_end = dt_end.astimezone()
+            event = {
+                "start-date": dt_start.strftime(dt_format),
+                "end-date": dt_end.strftime(dt_format),
+                "machine": device_name,
+                "event-type": l4_event["event-type"],
+                "organization": l4_event.get("organization", "fairshare"),
+            }
+            calendar_data.append(event)
+        calendar_data.sort(
+            key=lambda item: datetime.datetime.strptime(item["start-date"], dt_format)
+        )
+        return calendar_data
+
     @property
     def backend_info(self) -> Optional[BackendInfo]:
         if self._backend_info is None and not self._MACHINE_DEBUG:
@@ -442,7 +546,7 @@ class QuantinuumBackend(Backend):
     @property
     def _gate_set(self) -> Set[OpType]:
         return (
-            _GATE_SET
+            _ALL_GATES
             if self._MACHINE_DEBUG
             else cast(BackendInfo, self.backend_info).gate_set
         )
@@ -506,13 +610,23 @@ class QuantinuumBackend(Backend):
             scratch_reg_resize_pass(),
         ]
         squash = auto_squash_pass({OpType.PhasedX, OpType.Rz})
+        target_2qb_gate = self.compilation_config.target_2qb_gate
+        assert target_2qb_gate is not None
         # use default (perfect fidelities) for supported gates
         fidelities: Dict[str, Any] = {}
-        # If ZZPhase is available we should prefer it to ZZMax.
-        if OpType.ZZPhase in self._gate_set:
-            fidelities["ZZPhase_fidelity"] = lambda x: 1.0
+        if target_2qb_gate == OpType.TK2:
+            decomposition_passes = []
+        elif target_2qb_gate == OpType.ZZPhase:
+            decomposition_passes = [
+                NormaliseTK2(),
+                DecomposeTK2(ZZPhase_fidelity=lambda x: 1.0),
+            ]
+        elif target_2qb_gate == OpType.ZZMax:
+            decomposition_passes = [NormaliseTK2(), DecomposeTK2(ZZMax_fidelity=1.0)]
         else:
-            fidelities["ZZMax_fidelity"] = 1.0
+            raise ValueError(
+                f"Unrecognized target 2-qubit gate: {target_2qb_gate.name}"
+            )
         # If you make changes to the default_compilation_pass,
         # then please update this page accordingly
         # https://tket.quantinuum.com/extensions/pytket-quantinuum/index.html#default-compilation
@@ -520,11 +634,10 @@ class QuantinuumBackend(Backend):
         if optimisation_level == 0:
             passlist.append(self.rebase_pass())
         elif optimisation_level == 1:
+            passlist.append(SynthesiseTK())
+            passlist.extend(decomposition_passes)
             passlist.extend(
                 [
-                    SynthesiseTK(),
-                    NormaliseTK2(),
-                    DecomposeTK2(**fidelities),
                     self.rebase_pass(),
                     ZZPhaseToRz(),
                     RemoveRedundancies(),
@@ -533,11 +646,10 @@ class QuantinuumBackend(Backend):
                 ]
             )
         else:
+            passlist.append(FullPeepholeOptimise(target_2qb_gate=OpType.TK2))
+            passlist.extend(decomposition_passes)
             passlist.extend(
                 [
-                    FullPeepholeOptimise(target_2qb_gate=OpType.TK2),
-                    NormaliseTK2(),
-                    DecomposeTK2(**fidelities),
                     self.rebase_pass(),
                     RemoveRedundancies(),
                     squash,
@@ -861,7 +973,13 @@ class QuantinuumBackend(Backend):
             else:
                 results_selection = []
                 if language == Language.QASM:
-                    quantinuum_circ = circuit_to_qasm_str(c0, header="hqslib1")
+                    quantinuum_circ = circuit_to_qasm_str(
+                        c0,
+                        header="hqslib1",
+                        maxwidth=self.backend_info.misc["cl_reg_width"]
+                        if self.backend_info
+                        else 32,
+                    )
                     used_scratch_regs = _used_scratch_registers(quantinuum_circ)
                     for name, count in Counter(
                         bit.reg_name
@@ -1345,3 +1463,18 @@ def _parse_status(response: Dict) -> CircuitStatus:
     }
     message = json.dumps(msgdict)
     return CircuitStatus(_STATUS_MAP[h_status], message)
+
+
+def _convert_datetime_string(datetime_string: str) -> datetime.datetime:
+    year, month, day = list(map(int, datetime_string[:10].split("-")))
+    hour, minute, second = list(map(int, datetime_string[11:].split(":")))
+    dt = datetime.datetime(
+        year=year,
+        month=month,
+        day=day,
+        hour=hour,
+        minute=minute,
+        second=second,
+        tzinfo=datetime.timezone.utc,
+    )
+    return dt