+++ docs/changelog.rst
@@ -4,8 +4,18 @@ Changelog
 Unreleased
 ----------
 
+* Add method to enable visibility of 
+Quantinuum H-Series operational calendar.
+* Update pytket_pecos version requirement to 0.1.18.
+* Support TK2 as native gate.
+
+0.30.0 (February 2024)
+----------------------
+
 * Make pytket-qir an automatic dependency.
+* Update pytket version requirement to 1.25.
 * Update pytket-qir version requirement to 0.5.
+* Update pytket_pecos version requirement to 0.1.17.
 
 0.29.0 (January 2024)
 ---------------------